refactor(qna): 고객문의 api 조회시 nickname 추가
