+++ src/api/service/qna/getQnaListData.ts
@@ -9,12 +9,25 @@ export interface Qna {
   is_answer: boolean;
   answer: string;
   answer_at: string;
+  profiles: {
+    nickname: string;
+  };
 }
 
 export type QnaList = Qna[];
 
 export const getQnaList = async (): Promise<QnaList | null> => {
-  const { data, error } = await supabase.from("customer_support").select("*");
+  const { data, error } = await supabase
+    .from("customer_support")
+    .select(
+      `
+      *,
+      profiles (
+        nickname
+      )
+    `
+    )
+    .order("created_at", { ascending: false });
   if (error || !data) {
     console.log("고객 문의 센터 조회 api 실패:", error);
   }
